requirements.txt
gtts==2.5.4
python-telegram-bot==22.0
